Add type alias and fix some tests

+++ storage/series/series.go
@@ -152,7 +152,7 @@ func (s *dbSeries) updateBlocksWithLock() updateBlocksResult {
 	)
 	var start time.Time
 	for startNano, currBlock := range s.blocks.AllBlocks() {
-		start = time.Unix(0, startNano)
+		start = time.Unix(0, int64(startNano))
 		if start.Before(expireCutoff) {
 			s.blocks.RemoveBlockAt(start)
 			currBlock.Close()
@@ -334,7 +334,7 @@ func (s *dbSeries) FetchBlocksMetadata(
 
 	var t time.Time
 	for tNano, b := range blocks {
-		t = time.Unix(0, tNano)
+		t = time.Unix(0, int64(tNano))
 		if !start.Before(t.Add(blockSize)) || !t.Before(end) {
 			continue
 		}
@@ -448,7 +448,7 @@ func (s *dbSeries) Bootstrap(blocks block.DatabaseSeriesBlocks) error {
 		min, _ := s.buffer.MinMax()
 		var t time.Time
 		for tNano, block := range blocks.AllBlocks() {
-			t = time.Unix(0, tNano)
+			t = time.Unix(0, int64(tNano))
 			if !t.Before(min) {
 				if err := s.buffer.Bootstrap(block); err != nil {
 					multiErr = multiErr.Add(s.newBootstrapBlockError(block, err))
